Revert "Implement configurable pre filters"
This reverts commit fbb9f079c297fee62740d2114a0cbb7915ce167d.
