+++ wagtail_wordpress_import/bleach.py
@@ -1,5 +1,27 @@
 from bs4 import BeautifulSoup as bs4
-from django.utils.module_loading import import_string
+from wagtail_wordpress_import.constants import (
+    ALLOWED_TAGS,
+    ALLOWED_STYLES,
+    ALLOWED_ATTRIBUTES,
+    FILTER_MAPPING,
+    HTML_TAGS,
+)
+
+
+from bleach.sanitizer import Cleaner
+
+
+def bleach_clean(value):
+    """
+    Clean up the raw html to be on the safe side.
+    Keeping all styles in place that we know of and care about.
+    See ALLOWED lists in wagtail-wordpress-import/wagtail_wordpress_import/constants.py
+    """
+
+    cleaned = Cleaner(
+        tags=ALLOWED_TAGS, attributes=ALLOWED_ATTRIBUTES, styles=ALLOWED_STYLES
+    )
+    return cleaned.clean(value)
 
 
 def reverse_styles_dict(mapping):
@@ -11,15 +33,15 @@ def reverse_styles_dict(mapping):
     """
     # note no ending `;` so we can split on it later
     {
-        'FONT-WEIGHT: bold': 'boldify',
-        'font-weight: bold': 'boldify',
+        'FONT-WEIGHT: bold': 'boldify', 
+        'font-weight: bold': 'boldify', 
         'font-weight: bold; color: #006600': 'boldify'
     }
     """
     return inverse
 
 
-def filter_fix_styles(html, options=None):
+def fix_styles(value):
     """
     This function uses the mapping of the style attribute for an element to break
     matching elements into one or more html tags.
@@ -30,28 +52,9 @@ def filter_fix_styles(html, options=None):
     later to decide if an element can have alignment in the richtext block
     e.g. "margin: 0pt 10px 0px 0pt; float: left;" maps to "leftfloat"
     e.g. "float: left; margin: 0em 1em 1em 0em;" maps to "leftfloat"
-
-    param: `options` NOT IMPLEMENTED
     """
-
-    CONF_HTML_TAGS = HTML_TAGS
-    if options and options["CONFIG"].get("HTML_TAGS"):
-        html_tags = import_string(options["CONFIG"]["HTML_TAGS"])
-        if callable(html_tags):
-            CONF_HTML_TAGS = html_tags()
-        else:
-            CONF_HTML_TAGS = html_tags
-
-    CONF_FILTER_MAPPING = FILTER_MAPPING
-    if options and options["CONFIG"].get("FILTER_MAPPING"):
-        filter_mapping = import_string(options["CONFIG"]["FILTER_MAPPING"])
-        if callable(filter_mapping):
-            CONF_FILTER_MAPPING = filter_mapping()
-        else:
-            CONF_FILTER_MAPPING = filter_mapping
-
-    soup = bs4(html, "html.parser")
-    search_styles = reverse_styles_dict(CONF_FILTER_MAPPING)
+    soup = bs4(value, "html.parser")
+    search_styles = reverse_styles_dict(FILTER_MAPPING)
 
     for style_string in search_styles:
         filter = search_styles[style_string]
@@ -62,7 +65,7 @@ def filter_fix_styles(html, options=None):
             # directly backk into the final returned soup
 
             try:
-                item_type = CONF_HTML_TAGS[item.name]
+                item_type = HTML_TAGS[item.name]
             except KeyError:
                 print("item.name = tag not found in HTML_TAGS")
                 continue
@@ -151,111 +154,4 @@ def filter_fix_styles(html, options=None):
         new_item.string = item.text
         item.replace_with(new_item)
 
-    fixed_html = str(soup)
-
-    return fixed_html
-
-
-HTML_TAGS = {
-    "address": "block",
-    "article": "block",
-    "aside": "block",
-    "blockquote": "block",
-    "canvas": "block",
-    "dd": "block",
-    "div": "block",
-    "dl": "block",
-    "dt": "block",
-    "fieldset": "block",
-    "figcaption": "block",
-    "figure": "block",
-    "footer": "block",
-    "form": "block",
-    "h1": "block",
-    "h2": "block",
-    "h3": "block",
-    "h4": "block",
-    "h5": "block",
-    "h6": "block",
-    "header": "block",
-    "hr": "block",
-    "li": "block",
-    "main": "block",
-    "nav": "block",
-    "noscript": "block",
-    "ol": "block",
-    "p": "block",
-    "pre": "block",
-    "section": "block",
-    "table": "block",
-    "tfoot": "block",
-    "ul": "block",
-    "video": "block",
-    "a": "inline",
-    "abbr": "inline",
-    "acronym": "inline",
-    "b": "inline",
-    "bdo": "inline",
-    "big": "inline",
-    "br": "inline",
-    "button": "inline",
-    "center": "inline",  # not stricty allowed but here for later styling
-    "cite": "inline",
-    "code": "inline",
-    "dfn": "inline",
-    "em": "inline",
-    "i": "inline",
-    "img": "inline",
-    "input": "inline",
-    "kbd": "inline",
-    "label": "inline",
-    "map": "inline",
-    "object": "inline",
-    "output": "inline",
-    "q": "inline",
-    "samp": "inline",
-    "script": "inline",
-    "select": "inline",
-    "small": "inline",
-    "span": "inline",
-    "strong": "inline",
-    "sub": "inline",
-    "sup": "inline",
-    "textarea": "inline",
-    "time": "inline",
-    "tt": "inline",
-    "var": "inline",
-}
-
-FILTER_MAPPING = {
-    "bold": [
-        # transform to <b></b>
-        "font-weight:bold;",
-    ],
-    "italic": [
-        # transform to <i></i>
-        "font-style:italic;",
-    ],
-    "bold-italic": [
-        # transform to <b><i></i></b>
-        "font-style:italic; font-weight:bold;",
-        "font-weight:bold; font-style:italic;",
-    ],
-    "center": [
-        # add class align-center
-        "text-align:center",
-    ],
-    "leftfloat": [
-        # add class float-left
-        "float:left;"
-    ],
-    "rightfloat": [
-        # add class float-right
-        "float:right;",
-    ],
-    "remove": [
-        # remove style tag completely
-        "font-weight:400;",
-        "font-weight:normal;",
-    ],
-}
+    return str(soup)
